refactor: Card 객체 인스턴스를 미리 만들어 관리하기 위한 Cards 오브젝트 추출
