+++ src/main/kotlin/camp/nextstep/blackjack/card/Cards.kt
@@ -1,0 +1,25 @@
+package camp.nextstep.blackjack.card
+
+object Cards {
+
+    private val CARDS = CardSuit.values()
+        .associateWith { CardNumber.values() }
+        .mapValues { (suit, numbers) -> numbers.map { suit.with(it) } }
+        .values
+        .flatten()
+        .associateBy { key(it.suit, it.number) }
+
+    fun ofCombinations(): Set<Card> {
+        return CARDS.values.toSet()
+    }
+
+    fun of(suit: CardSuit, number: CardNumber): Card {
+        return requireNotNull(CARDS[key(suit, number)]) { "잘못된 카드 모양과 숫자입니다." }
+    }
+
+    private fun key(suit: CardSuit, number: CardNumber): String {
+        return suit.name + number.name
+    }
+
+    private fun CardSuit.with(number: CardNumber) = Card(this, number)
+}
